+++ Telemonitor/main.py
@@ -4,15 +4,14 @@ import logging
 from aiogram import Bot, Dispatcher, executor, types
 from aiogram.utils.markdown import bold, code
 
-from Telemonitor import helpers as h, \
-    __version__
+from Telemonitor import helpers as h, __version__
 
 
 def run():
-    # Initialization
     chdir(path.dirname(__file__))
     h.init_logger()
     logger = logging.getLogger('TM.Main')
+    logger.info("Bot is starting")
 
     cfg = h.TM_Config().get()
     wls = h.TM_Whitelist()
@@ -28,14 +27,13 @@ def run():
     async def __command_start(message: types.Message):
         if wls.is_whitelisted(message.from_user.id):
             await message.reply(
-                f'{bold("Welcome to the Telemonitor control panel.")}\nVersion: {code(__version__)}\n',
+                f'{bold("Welcome to the Telemonitor control panel.")}\n\nVersion: {code(__version__)}\n',
                 reply=False,
                 parse_mode=h.PARSE_MODE,
                 reply_markup=ikb.get_keyboard()
             )
 
     print('Bot is starting')
-    logger.info("Bot is starting")
     executor.start_polling(
         dp,
         skip_updates=True,